[Refactor] 프로필 모달 개선
