+++ src/components/common/Header/HeaderProfileItem.tsx
@@ -1,47 +1,40 @@
 'use client'
 
 import clsx from 'clsx'
-import LoginModal from '@/modals/LoginModal'
-import Portal from '@/portal/Portal'
-import { useDispatch, useSelector } from 'react-redux'
-import { RootState } from '@/redux/store'
+import { useDispatch } from 'react-redux'
 import { setIsOpen } from '@/redux/features/profileModalSlice'
+import Link from 'next/link'
 
 function HeaderProfileItem({ section, chosenItem, setChosenItem }) {
   const dispatch = useDispatch()
-  const isOpen = useSelector((state: RootState) => state.profile.isOpen)
 
   function showModal() {
     dispatch(setIsOpen(true))
   }
 
-  function closeModal() {
-    dispatch(setIsOpen(false))
-  }
   function handleClick() {
     if (section.key === 'signup') {
       showModal()
     }
     setChosenItem(section.key)
   }
+
   return (
-    <>
-      <li
-        onClick={handleClick}
-        className={clsx(
-          'cursor-pointer px-4 py-3 text-sm transition-transform hover:rounded-lg hover:bg-gray-200',
-          {
-            'font-bold': chosenItem === section.key,
-          },
-        )}>
-        {section.title}
-      </li>
-      {isOpen && (
-        <Portal>
-          <LoginModal onClose={closeModal} />
-        </Portal>
+    <li
+      className={clsx(
+        'cursor-pointer px-4 py-3 text-sm transition-transform hover:rounded-lg hover:bg-gray-200',
+        {
+          'font-bold': chosenItem === section.key,
+        },
+      )}>
+      {section.key === 'profile' ? (
+        <Link href='/profile' onClick={handleClick}>
+          {section.title}
+        </Link>
+      ) : (
+        <li onClick={handleClick}>{section.title}</li>
       )}
-    </>
+    </li>
   )
 }
